Add TaskController and schedule building
